chore: clean up wordsgrid component
