+++ src/app/GameBoard/WordsGrid/index.tsx
@@ -8,18 +8,11 @@ import {
   useTrail,
   useChain,
   easings,
-  SpringValue,
 } from '@react-spring/web';
 import { makeSolutionSets } from '../SolutionDisplay';
+import LastPlayedWord from './LastPlayedWord';
 
-const WordsGridContainer = styled('div', {
-  width: '324px',
-  height: '324px',
-  display: 'flex',
-  flexDirection: 'column',
-});
-
-const TileWrapper = styled('div', {
+const WordWrapper = styled('div', {
   display: 'flex',
   margin: '12px 0px 0px',
 });
@@ -29,10 +22,10 @@ const Badge = styled(animated.div, {
   justifyContent: 'center',
   alignItems: 'center',
   color: 'black',
-  height: '24px',
-  width: '24px',
+  height: '32px',
+  width: '32px',
   marginLeft: '8px',
-  border: '1.5px solid black',
+  border: '2px solid black',
   borderRadius: '50%',
 });
 
@@ -41,7 +34,7 @@ const BadgeContents = styled(animated.div, {
   justifyContent: 'center',
   alignItems: 'center',
   fontFamily: 'Helvetica',
-  fontSize: '11px',
+  fontSize: '12px',
   color: 'black',
 });
 
@@ -102,80 +95,49 @@ const WordsGrid: React.FC<Props> = ({ playedWords, solutionSet }) => {
   const solutionSets = makeSolutionSets(playedWords, solutionSet);
 
   useChain([trailsRef, springRef]);
-  return (
-    <WordsGridContainer className="WGContainer">
-      {makeWordsGrid(playedWords).map((word: string[], wordIdx: number) => {
-        const isPendingWord = !!word[0] && !word[1];
-        const isLastPlayedWord = wordIdx === playedWords.length - 1;
-        return (
-          <animated.div
-            style={{
-              display: 'flex',
-              flexDirection: 'column',
-            }}
-            key={wordIdx}
-          >
-            <TileWrapper>
-              {isLastPlayedWord ? (
-                <AnimatedWord
-                  word={word}
-                  trails={trails}
-                  isLastTurn={playedWords.length >= 3}
-                />
-              ) : (
-                <div style={{ display: 'flex', alignItems: 'center' }}>
-                  {word.map((letter, letterIdx) => (
-                    <Tile
-                      key={letterIdx}
-                      letter={letter}
-                      isPendingWord={isPendingWord}
-                      isAnchorTile={[0].includes(letterIdx)}
-                    />
-                  ))}
-                  {isPendingWord && (
-                    <Badge style={{ ...spring }}>
-                      <BadgeContents>
-                        {solutionSets[wordIdx]?.size ?? '!'}
-                      </BadgeContents>
-                    </Badge>
-                  )}
-                </div>
-              )}
-            </TileWrapper>
-            {wordIdx !== 3 && <GuideLine />}
-          </animated.div>
-        );
-      })}
-    </WordsGridContainer>
-  );
-};
 
-interface WordProps {
-  word: string[];
-  trails: {
-    transform: SpringValue<string>;
-  }[];
-  isLastTurn: boolean;
-}
-// animated on mount
-const AnimatedWord: React.FC<WordProps> = ({ word, trails, isLastTurn }) => {
-  return (
-    <>
-      {trails.map((trail, idx) => {
-        // first and last letters in array will be right and left bounds of board
-        const isAnchorTile = isLastTurn ? idx === 0 : [0, 4].includes(idx);
-        return (
-          <Tile
-            key={idx}
-            letter={word[idx]}
-            isPendingWord={false}
-            isAnchorTile={isAnchorTile}
-            spring={trail}
-          />
-        );
-      })}
-    </>
-  );
+  return makeWordsGrid(playedWords).map((word: string[], wordIdx: number) => {
+    const isPendingWord = !!word[0] && !word[1];
+    const isLastPlayedWord = wordIdx === playedWords.length - 1;
+    return (
+      <animated.div
+        style={{
+          display: 'flex',
+          flexDirection: 'column',
+        }}
+        key={wordIdx}
+      >
+        <WordWrapper>
+          {isLastPlayedWord ? (
+            <LastPlayedWord
+              word={word}
+              trails={trails}
+              isLastTurn={playedWords.length >= 3}
+            />
+          ) : (
+            <div style={{ display: 'flex', alignItems: 'center' }}>
+              {word.map((letter, letterIdx) => (
+                <Tile
+                  key={letterIdx}
+                  letter={letter}
+                  isPendingWord={isPendingWord}
+                  isAnchorTile={[0].includes(letterIdx)}
+                />
+              ))}
+              {isPendingWord && (
+                <Badge style={{ ...spring }}>
+                  <BadgeContents>
+                    {solutionSets[wordIdx]?.size ?? '0'}
+                  </BadgeContents>
+                </Badge>
+              )}
+            </div>
+          )}
+        </WordWrapper>
+        {wordIdx !== 3 && <GuideLine />}
+      </animated.div>
+    );
+  });
 };
 
 export default WordsGrid;
